+++ sec_edgar_client/__init__.py
@@ -1,13 +1,15 @@
 from .central_index_key import CIKRepositoryInterface, LocalFileCIKRepository
 from .client import SECClient, UserAgent
-from .parser import BalanceSnapshot, Reports
+from .entities import Balance, Income, Reports, Year
 
 __all__ = (
     "SECClient",
     "UserAgent",
 
     "Reports",
-    "BalanceSnapshot",
+    "Balance",
+    "Income",
+    "Year",
 
     "CIKRepositoryInterface",
     "LocalFileCIKRepository",