+++ sec_edgar_client/parser.py
@@ -3,26 +3,11 @@ from typing import Iterable, Mapping
 
 import attr
 
-from .utils import get_trimmed_to_same_len
+from .entities import Balance, Income, Reports, Year
 
 __all__ = (
     "SECResponseParser",
-
-    "Reports",
-    "BalanceSnapshot",
 )
-
-
-@attr.s(auto_attribs=True, slots=True, frozen=True)
-class BalanceSnapshot:
-    assets: int
-    equity: int
-
-
-@attr.s(auto_attribs=True, slots=True, frozen=True)
-class Reports:
-    balance: tuple[BalanceSnapshot, ...]
-    reported_at: tuple[date, ...]
 
 
 @attr.s(auto_attribs=True, slots=True, frozen=True)
@@ -30,58 +15,97 @@ class SECResponseParser:
     response: Mapping
 
     def parse_reports(self) -> Reports:
-        # pylint: disable=unbalanced-tuple-unpacking
-        assets, equity = get_trimmed_to_same_len(
-            _get_statements(self.response, key="Assets"),
-            _get_statements(self.response, key="StockholdersEquity"),
-        )
-
-        balance = tuple(
-            BalanceSnapshot(asset_value, equity_value)
-            for asset_value, equity_value in zip(
-                assets.values(),
-                equity.values(),
-            )
-        )
         return Reports(
-            balance,
-            reported_at=tuple(assets.keys()),
+            balance=self._parse_balance(),
+            income=self._parse_income(),
         )
 
+    def _parse_balance(self) -> Balance:
+        return Balance(
+            assets=self._get_statements(key="Assets"),
+            equity=self._get_statements(key="StockholdersEquity"),
+        )
 
-def _get_statements(data: Mapping, key: str) -> Mapping[date, int]:
-    statements = data["facts"]["us-gaap"][key]["units"]["USD"]
-    annual_statements = _get_annual_statements(statements)
-    sorted_annual_statements = _get_sorted_by_date(annual_statements)
-    return _get_separated_date_and_value(sorted_annual_statements)
+    def _parse_income(self) -> Income:
+        revenue = self._get_statements(key="Revenues")
+        revenue_from_contracts = self._get_statements(
+            key="RevenueFromContractWithCustomerExcludingAssessedTax",
+        )
+        return Income(
+            revenue=revenue_from_contracts | revenue,
+            gross_profit=self._get_statements(key="GrossProfit"),
+            operating_income=self._get_statements(key="OperatingIncomeLoss"),
+            net_income=self._get_statements(key="NetIncomeLoss"),
+            research_and_development=self._get_statements(
+                key="ResearchAndDevelopmentExpense",
+            ),
+            selling_and_marketing=self._get_statements(
+                key="SellingAndMarketingExpense",
+            ),
+            general_and_administrative=self._get_statements(
+                key="GeneralAndAdministrativeExpense",
+            ),
+        )
+
+    def _get_statements(self, key: str) -> dict[Year, int]:
+        statements = self.response["facts"]["us-gaap"][key]["units"]["USD"]
+        annual_statements = _get_annual_statements(statements)
+        sorted_annual_statements = _get_sorted_statements_by_actuality(
+            annual_statements,
+        )
+        year_to_statement = _get_separated_year_and_value(
+            sorted_annual_statements,
+        )
+        return {
+            key: year_to_statement[key]
+            for key in sorted(year_to_statement)
+        }
 
 
-def _get_sorted_by_date(statements: Iterable[Mapping]) -> Iterable[Mapping]:
-    return sorted(
-        statements,
-        key=lambda statement: (
-            date.fromisoformat(statement["end"]), statement["fy"],
-        ),
-    )
+def _get_annual_statements(statements: Iterable[Mapping]) -> Iterable[dict]:
+    for statement in statements:
+        if statement["form"] == "10-K":
+            frame = statement.get("frame")
+            if frame is None or _is_year_frame(frame):
+                yield statement
 
 
-def _get_annual_statements(statements: Iterable[Mapping]) -> Iterable[Mapping]:
-    return (
-        statement
-        for statement in statements
-        if statement["form"] == "10-K" and "frame" not in statement
-    )
-
-
-def _get_separated_date_and_value(
+def _get_separated_year_and_value(
     statements: Iterable[Mapping],
-) -> dict[date, int]:
-    date_to_statement_value = {}
+) -> dict[Year, int]:
+    year_to_value = {}
 
     for statement in statements:
-        period_end = date.fromisoformat(statement["end"])
+        frame = statement.get("frame")
+        value = statement["val"]
+        if frame is None:
+            year = statement["fy"]
+        else:
+            year = _get_year_from_frame(statement["frame"])
 
-        if period_end not in date_to_statement_value:
-            date_to_statement_value[period_end] = statement["val"]
+        if year not in year_to_value:
+            year_to_value[year] = value
 
-    return date_to_statement_value
+    return year_to_value
+
+
+def _is_year_frame(value: str) -> bool:
+    without_prefix = value.removeprefix("CY")
+    frame = without_prefix[4:]  # remove year
+    return frame in {"", "Q4I"}
+
+
+def _get_year_from_frame(frame_value: str) -> Year:
+    return int(frame_value.removeprefix("CY")[:4])
+
+
+def _get_sorted_statements_by_actuality(
+    statements: Iterable[Mapping],
+) -> Iterable[Mapping]:
+    return sorted(statements, key=_sort_actuality_key, reverse=True)
+
+
+def _sort_actuality_key(statement: Mapping) -> tuple[date, date]:
+    period_end = date.fromisoformat(statement["end"])
+    filled_at = date.fromisoformat(statement["filed"])
+    return filled_at, period_end